Deduplicate container setup by adding a function to "docker exec"
This function just does the 3 lines that were done before,
make a cmd, log it, run it. We just don't have to write it out
each time, and can't forget one of the lines.

By declaring it each time around the for loop, `i` is recaptured
each time and we don't have to pass it to the function every time.

Simplified example:
```
>>> for i in range(3):
...     def foo():
...             return i
...     print(foo())
...
0
1
2
```