Update _review.md
Re-ordered questions so that they make more logical sense.

In next-steps.MD, there is reference to what the learner should now know. This needs to be added here, to the review section. In the template file that was initially filled out, this should be there. So the review section starts with 'you now know how ...', i.e., referring back to the initial LOs and is then followed by questions to test the learning.