Memory Tagged (MTE) Allocator Learning Path
This path builds on https://learn.arm.com/learning-paths/cross-platform/dynamic-memory-allocator/
by using the Memory Tagging Extension (MTE) to prevent common memory problems in the
memory allocator.

Full source is included in this path so readers don't have to copy it from the
previous path, however they will need to understand the basic design of the
allocator to understand the changes needed to support MTE.

So the full "journey" might be:
* The memory allocator path (https://learn.arm.com/learning-paths/cross-platform/dynamic-memory-allocator/)
* The stack overflows path (https://learn.arm.com/learning-paths/servers-and-cloud-computing/exploiting-stack-buffer-overflow-aarch64/)
* The intro to MTE path (https://learn.arm.com/learning-paths/smartphones-and-mobile/mte/)
* This memory tagged allocator path

The binary runs under qemu-user, so it is possible to follow the path
on current AArch64 hardware or non-AArch64 hardware.

I'm sure we can tie this to other MTE resources but I've left the links
empty as you'll probably have a better idea of what to put there.