Add a learning path on modern library develomment in C++.
This is the first commit, adding the first chapters of the learning
path. Other commits to come will add chapters on:
 - adding documentation
 - performance measurement and improvement
 - adding support for new instructions
 - deployment

At the end of each chapter, the markdown offers readers to download an
archive with the Matrix library project (in the state in which it is at
the end of the chapter). This commit does *not* contain the archives.  A
decision should be made about either adding those archive directly in
the arm-learning-paths git repository, which is not very elegant on top
of being error prone. A prefered solution (partly implemented in this
commit) would be to generate those artifacts into
(static/artifacts/learning-path-name/) with a hook whenever the learning
paths are updated and pushed onto to the server.