Make hugo-server.sh more useful.
This patch improves hugo-server.sh to make it useable as-is on a number
of common platforms, instead of being merely an illustration of the
steps to do.

If this is thought to be a good idea, then I can update the
documentation (i.e. "Learning Path setup > Enable the search box on the
home page") accordingly.