Update Example_application.md
Updated comment in code snippet in Example_application.md to address review comment: In a code snippet in Example_application.md a comment reads "Sleep for 1 second (1000 milliseconds)" whereas the program code two lines below reads "Thread.sleep(250);". The correction would be to change the comment to read "Sleep for 1/4 second (250 milliseconds)". This would then be consistent with the description above, which states that the example prints out four times a second.